requirements.txt
numpy
pandas
Pillow==10.3
plotly
pytesseract
pywin32
opencv-python